+++ resources/assets/js/datatable/api.ts
@@ -65,3 +65,54 @@ export function fnGetSelection(tableId: string) {
 export function getApiFromSettings(settings: Settings) {
 	return new $.fn.dataTable.Api(settings as any);
 }
+
+/**
+ * Wordt gebruikt in gesprekken.
+ *
+ * @param {jQuery} $table
+ */
+function fnAutoScroll($table: JQuery) {
+	const $scroll = $table.parent();
+	if ($scroll.hasClass('dataTables_scrollBody')) {
+		// autoscroll if already on bottom
+		if ($scroll.scrollTop()! + $scroll.innerHeight()! >= $scroll[0].scrollHeight - 20) {
+			// check before draw and scroll after
+			window.setTimeout(() => {
+				$scroll.animate({
+					scrollTop: $scroll[0].scrollHeight,
+				}, 800);
+			}, 200);
+		}
+	}
+}
+
+export const fnGetLastUpdate = ($table: JQuery) => () => Number($table.data('lastupdate'));
+export const fnSetLastUpdate = ($table: JQuery) => (lastUpdate: number) => $table.data('lastupdate', lastUpdate);
+/**
+ * Called after ajax load complete.
+ *
+ * @returns object
+ * @param {jQuery} $table
+ */
+export const fnAjaxUpdateCallback = ($table: JQuery) => (json: DatatableResponse) => {
+	fnSetLastUpdate($table)(json.lastUpdate);
+	const tableConfig = $table.DataTable();
+
+	if (json.autoUpdate) {
+		const timeout = parseInt(json.autoUpdate, 10);
+		if (!isNaN(timeout) && timeout < 600000) { // max 10 min
+			setTimeout(() => {
+				$.post(tableConfig.ajax.url(), {
+					lastUpdate: fnGetLastUpdate($table),
+				}, (data) => {
+					fnUpdateDataTable($table.attr('id')!, data);
+					fnAjaxUpdateCallback($table)(data);
+				});
+			}, timeout);
+		}
+	}
+
+	fnAutoScroll($table);
+
+	return json.data;
+};
